chore: split tests in a new objects.test.ts
